third_party/blink/web_tests/svg/dynamic-updates/SVGRadialGradientElement-svgdom-r-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'r' property of the SVGRadialGradientElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS radialGradientElement.r.baseVal.valueAsString is "125%"
PASS radialGradientElement.r.baseVal.valueAsString is "50%"
PASS successfullyParsed is true

TEST COMPLETE
